+++ CHANGELOG.rst
@@ -4,7 +4,7 @@ Changelog
 0.1.1 (upcoming)
 ----------------
 
-* ...
+* Fixed loading .ipuz puzzles that do not have a solution element.
 
 0.1 (August 17, 2014)
 ---------------------