+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ Changelog
 
 * Fixed loading .ipuz puzzles that do not have a solution element.
 * Fixed loading .ipuz puzzles that do not have Across and/or Down clues.
+* Fixed writing .ipuz puzzles to only include "block" and/or "empty" when specified.
 
 0.1 (August 17, 2014)
 ---------------------