+++ CHANGELOG.rst
@@ -4,8 +4,8 @@ Changelog
 0.1.1 (upcoming)
 ----------------
 
-* Fixed loading .ipuz puzzles that do not have a solution element.
-* Fixed loading .ipuz puzzles that do not have Across and/or Down clues.
+* Fixed reading .ipuz puzzles that do not have a solution element.
+* Fixed reading .ipuz puzzles that do not have Across and/or Down clues.
 * Fixed writing .ipuz puzzles to only include "block" and/or "empty" when specified.
 
 0.1 (August 17, 2014)